+++ internal/topic/topicreaderinternal/batcher.go
@@ -12,11 +12,12 @@ import (
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xsync"
 )
 
-var errRemoveUnexpectedWaiter = xerrors.Wrap(errors.New("ydb: remove unexpected waiter"))
+var errBatcherPopConcurency = xerrors.Wrap(errors.New("ydb: batch pop concurency, internal state error"))
 
 type batcher struct {
-	waiterID uint64
-	closeErr error
+	popInFlight    int64
+	closeErr       error
+	hasNewMessages empty.Chan
 
 	m xsync.Mutex
 
@@ -24,13 +25,13 @@ type batcher struct {
 	closed                                        bool
 	closeChan                                     empty.Chan
 	messages                                      batcherMessagesMap
-	waiters                                       []batcherWaiter
 }
 
 func newBatcher() *batcher {
 	return &batcher{
-		messages:  make(batcherMessagesMap),
-		closeChan: make(empty.Chan),
+		messages:       make(batcherMessagesMap),
+		closeChan:      make(empty.Chan),
+		hasNewMessages: make(empty.Chan, 1),
 	}
 }
 
@@ -48,14 +49,20 @@ func (b *batcher) Close(err error) error {
 	return nil
 }
 
-func (b *batcher) PushBatch(batch *PublicBatch) error {
+func (b *batcher) PushBatches(batches ...*PublicBatch) error {
 	b.m.Lock()
 	defer b.m.Unlock()
 	if b.closed {
 		return xerrors.WithStackTrace(fmt.Errorf("ydb: push batch to closed batcher :%w", b.closeErr))
 	}
 
-	return b.addNeedLock(batch.commitRange.partitionSession, newBatcherItemBatch(batch))
+	for _, batch := range batches {
+		if err := b.addNeedLock(batch.commitRange.partitionSession, newBatcherItemBatch(batch)); err != nil {
+			return err
+		}
+	}
+
+	return nil
 }
 
 func (b *batcher) PushRawMessage(session *partitionSession, m rawtopicreader.ServerMessage) error {
@@ -83,7 +90,7 @@ func (b *batcher) addNeedLock(session *partitionSession, item batcherMessageOrde
 
 	b.messages[session] = currentItems
 
-	b.fireWaitersNeedLock()
+	b.notifyAboutNewMessages()
 
 	return nil
 }
@@ -143,117 +150,66 @@ func (o batcherGetOptions) splitBatch(batch *PublicBatch) (head, rest *PublicBat
 }
 
 func (b *batcher) Pop(ctx context.Context, opts batcherGetOptions) (_ batcherMessageOrderItem, err error) {
+	counter := atomic.AddInt64(&b.popInFlight, 1)
+	defer atomic.AddInt64(&b.popInFlight, -1)
+
+	if counter != 1 {
+		return batcherMessageOrderItem{}, xerrors.WithStackTrace(errBatcherPopConcurency)
+	}
+
 	if err = ctx.Err(); err != nil {
 		return batcherMessageOrderItem{}, err
 	}
-	var findRes batcherResultCandidate
-	var closed bool
 
-	var waiter batcherWaiter
-	defer func() {
-		removeWaiterErr := b.RemoveWaiter(waiter)
-		if err == nil {
-			err = removeWaiterErr
-		}
-	}()
-
-	b.m.WithLock(func() {
-		closed = b.closed
-		if closed {
-			return
-		}
-
-		findRes = b.findNeedLock(0, []batcherWaiter{{Options: opts}})
-		if findRes.Ok {
-			b.applyNeedLock(findRes)
-			return
-		}
-
-		waiter = b.createWaiterNeedLock(opts)
-	})
-	if closed {
-		return batcherMessageOrderItem{},
-			xerrors.WithStackTrace(fmt.Errorf("ydb: try pop messages from closed batcher: %w", b.closeErr))
-	}
-	if findRes.Ok {
-		return findRes.Result, nil
-	}
-
-	select {
-	case batch := <-waiter.Result:
-		return batch, nil
-	case <-b.closeChan:
-		return batcherMessageOrderItem{}, xerrors.WithStackTrace(fmt.Errorf("ydb: batcher closed: %w", b.closeErr))
-	case <-ctx.Done():
-		return batcherMessageOrderItem{}, ctx.Err()
-	}
-}
-
-func (b *batcher) createWaiterNeedLock(opts batcherGetOptions) batcherWaiter {
-	waiter := batcherWaiter{
-		Options:          opts,
-		Result:           make(chan batcherMessageOrderItem),
-		finishWaitSignal: make(empty.Chan),
-	}
-
-	// defend from overflow
-	for waiter.ID == 0 {
-		waiter.ID = atomic.AddUint64(&b.waiterID, 1)
-	}
-
-	b.waiters = append(b.waiters, waiter)
-
-	return waiter
-}
-
-func (b *batcher) RemoveWaiter(waiter batcherWaiter) error {
-	if waiter.ID == 0 {
-		return nil
-	}
-
-	close(waiter.finishWaitSignal)
-
-	return b.removeWaiterByID(waiter.ID)
-}
-
-func (b *batcher) removeWaiterByID(waiterID uint64) error {
-	b.m.Lock()
-	defer b.m.Unlock()
-
-	for i := 0; i < len(b.waiters); i++ {
-		if b.waiters[i].ID == waiterID {
-			b.removeWaiterByIndexNeedLock(i)
-			return nil
-		}
-	}
-
-	return xerrors.WithStackTrace(errRemoveUnexpectedWaiter)
-}
-
-func (b *batcher) fireWaitersNeedLock() {
-	startIndex := 0
 	for {
-		resCandidate := b.findNeedLock(startIndex, b.waiters)
-		if !resCandidate.Ok {
-			return
+		var findRes batcherResultCandidate
+		var closed bool
+
+		b.m.WithLock(func() {
+			closed = b.closed
+			if closed {
+				return
+			}
+
+			findRes = b.findNeedLock(opts)
+			if findRes.Ok {
+				b.applyNeedLock(findRes)
+				return
+			}
+		})
+		if closed {
+			return batcherMessageOrderItem{},
+				xerrors.WithStackTrace(fmt.Errorf("ydb: try pop messages from closed batcher: %w", b.closeErr))
+		}
+		if findRes.Ok {
+			return findRes.Result, nil
 		}
 
-		waiter := b.waiters[resCandidate.WaiterIndex]
-
+		// wait new messages for next iteration
 		select {
-		case waiter.Result <- resCandidate.Result:
-			// waiter receive the result, commit it
-			b.applyNeedLock(resCandidate)
-			return
-		case <-waiter.finishWaitSignal:
-			startIndex = resCandidate.WaiterIndex + 1
+		case <-b.hasNewMessages:
+			// new iteration
+		case <-b.closeChan:
+			return batcherMessageOrderItem{},
+				xerrors.WithStackTrace(
+					fmt.Errorf(
+						"ydb: batcher close while pop wait new messages: %w",
+						b.closeErr,
+					),
+				)
+		case <-ctx.Done():
+			return batcherMessageOrderItem{}, ctx.Err()
 		}
 	}
 }
 
-func (b *batcher) removeWaiterByIndexNeedLock(index int) {
-	copy(b.waiters[index:], b.waiters[index+1:])
-	b.waiters = b.waiters[:len(b.waiters)-1]
+func (b *batcher) notifyAboutNewMessages() {
+	select {
+	case b.hasNewMessages <- empty.Struct{}:
+		// sent signal
+	default:
+		// signal already in progress
+	}
 }
 
 type batcherResultCandidate struct {
@@ -268,20 +224,18 @@ func newBatcherResultCandidate(
 	key *partitionSession,
 	result batcherMessageOrderItem,
 	rest batcherMessageOrderItems,
-	waiterIndex int,
 	ok bool,
 ) batcherResultCandidate {
 	return batcherResultCandidate{
-		Key:         key,
-		Result:      result,
-		Rest:        rest,
-		WaiterIndex: waiterIndex,
-		Ok:          ok,
+		Key:    key,
+		Result: result,
+		Rest:   rest,
+		Ok:     ok,
 	}
 }
 
-func (b *batcher) findNeedLock(startIndex int, waiters []batcherWaiter) batcherResultCandidate {
-	if len(waiters) == 0 || len(b.messages) == 0 {
+func (b *batcher) findNeedLock(filter batcherGetOptions) batcherResultCandidate {
+	if len(b.messages) == 0 {
 		return batcherResultCandidate{}
 	}
 
@@ -293,31 +247,29 @@ func (b *batcher) findNeedLock(startIndex int, waiters []batcherWaiter) batcherR
 	for k, items := range b.messages {
 		head, rest, ok := rawMessageOpts.cutBatchItemsHead(items)
 		if ok {
-			return newBatcherResultCandidate(k, head, rest, len(waiters)-1, true)
+			return newBatcherResultCandidate(k, head, rest, true)
 		}
 
 		if needBatchResult {
-			for waiterIndex, waiter := range waiters[startIndex:] {
-				head, rest, ok = b.extractWaiterOptionsLeedLock(waiter).cutBatchItemsHead(items)
-				if !ok {
-					continue
-				}
-
-				needBatchResult = false
-				batchResult = newBatcherResultCandidate(k, head, rest, waiterIndex, true)
+			head, rest, ok = b.applyForceFlagToOptions(filter).cutBatchItemsHead(items)
+			if !ok {
+				continue
 			}
+
+			needBatchResult = false
+			batchResult = newBatcherResultCandidate(k, head, rest, true)
 		}
 	}
 
 	return batchResult
 }
 
-func (b *batcher) extractWaiterOptionsLeedLock(waiter batcherWaiter) batcherGetOptions {
+func (b *batcher) applyForceFlagToOptions(options batcherGetOptions) batcherGetOptions {
 	if !b.forceIgnoreMinRestrictionsOnNextMessagesBatch {
-		return waiter.Options
+		return options
 	}
 
-	res := waiter.Options
+	res := options
 	res.MinCount = 1
 	return res
 }
@@ -339,7 +291,7 @@ func (b *batcher) IgnoreMinRestrictionsOnNextPop() {
 	defer b.m.Unlock()
 
 	b.forceIgnoreMinRestrictionsOnNextMessagesBatch = true
-	b.fireWaitersNeedLock()
+	b.notifyAboutNewMessages()
 }
 
 type batcherMessagesMap map[*partitionSession]batcherMessageOrderItems
@@ -402,11 +354,4 @@ func (item *batcherMessageOrderItem) IsRawMessage() bool {
 
 func (item *batcherMessageOrderItem) IsEmpty() bool {
 	return item.RawMessage == nil && item.Batch.isEmpty()
-}
-
-type batcherWaiter struct {
-	ID               uint64
-	Options          batcherGetOptions
-	Result           chan batcherMessageOrderItem
-	finishWaitSignal empty.Chan
 }