+++ topic/reader_e2e_test.go
@@ -7,6 +7,7 @@ import (
 	"context"
 	"fmt"
 	"runtime/pprof"
+	"sync"
 	"sync/atomic"
 	"testing"
 	"time"
@@ -14,13 +15,16 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/ydb-platform/ydb-go-sdk/v3"
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/empty"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xcontext"
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xtest"
 	"github.com/ydb-platform/ydb-go-sdk/v3/table"
 	"github.com/ydb-platform/ydb-go-sdk/v3/table/types"
 	"github.com/ydb-platform/ydb-go-sdk/v3/topic/topicoptions"
 	"github.com/ydb-platform/ydb-go-sdk/v3/topic/topicreader"
 	"github.com/ydb-platform/ydb-go-sdk/v3/topic/topicsugar"
 	"github.com/ydb-platform/ydb-go-sdk/v3/topic/topictypes"
+	"github.com/ydb-platform/ydb-go-sdk/v3/trace"
 )
 
 func TestReadMessages(t *testing.T) {
@@ -71,7 +75,7 @@ func TestReadMessagesAndCommit(t *testing.T) {
 
 	sendCDCMessage(ctx, t, db)
 	sendCDCMessage(ctx, t, db)
-	reader = createReader(t, db)
+	reader = createFeedReader(t, db)
 
 	// read only no committed messages
 	for i := 0; i < 2; i++ {
@@ -115,9 +119,6 @@ func TestCDCFeedSendTopicPathSameAsSubscribed(t *testing.T) {
 }
 
 func TestTopicPath(t *testing.T) {
-	t.Skip("LOGBROKER-7625")
-	t.Skip("KIKIMR-14966")
-
 	ctx := testCtx(t)
 	db := connect(t)
 
@@ -129,13 +130,96 @@ func TestTopicPath(t *testing.T) {
 }
 
 func TestPartitionsBalanced(t *testing.T) {
-	ctx := testCtx(t)
-	db := connect(t)
-	topicPath := db.Name() + "/topic-" + t.Name()
+	xtest.TestManyTimes(t, func(t testing.TB) {
+		ctx := testCtx(t)
+		db := connect(t)
+		defer func() {
+			_ = db.Close(context.Background())
+		}()
+		topicPath := db.Name() + "/topic-" + t.Name()
 
-	_ = db.Topic().Drop(ctx, topicPath)
-	err := db.Topic().Create(ctx, topicPath, []topictypes.Codec{topictypes.CodecRaw})
-	require.NoError(t, err)
+		err := db.Topic().Drop(ctx, topicPath)
+		if err != nil {
+			require.True(t, ydb.IsOperationErrorSchemeError(err))
+		}
+
+		consumer := "test"
+		err = db.Topic().Create(ctx, topicPath, []topictypes.Codec{topictypes.CodecRaw},
+			topicoptions.CreateWithMinActivePartitions(2),
+			topicoptions.CreateWithPartitionCountLimit(2),
+			topicoptions.CreateWithConsumer(topictypes.Consumer{Name: consumer}),
+		)
+		require.NoError(t, err)
+
+		connectedPartitions := int32(0)
+		var handled int32
+
+		var sessionsMutex sync.Mutex
+		sessions := map[int64]bool{}
+
+		tracer := trace.Topic{
+			OnReaderPartitionReadStartResponse: func(startInfo trace.TopicReaderPartitionReadStartResponseStartInfo) func(doneInfo trace.TopicReaderPartitionReadStartResponseDoneInfo) { //nolint:lll
+				atomic.StoreInt32(&handled, 1)
+
+				atomic.AddInt32(&connectedPartitions, 1)
+				return nil
+			},
+			OnReaderPartitionReadStopResponse: func(startInfo trace.TopicReaderPartitionReadStopResponseStartInfo) func(doneInfo trace.TopicReaderPartitionReadStopResponseDoneInfo) { //nolint:lll
+				atomic.StoreInt32(&handled, 1)
+
+				sessionsMutex.Lock()
+				defer sessionsMutex.Unlock()
+				if sessions[startInfo.PartitionSessionID] {
+					return nil
+				}
+				sessions[startInfo.PartitionSessionID] = true
+
+				atomic.AddInt32(&connectedPartitions, -1)
+				return nil
+			},
+		}
+		firstReader, err := db.Topic().StartReader(consumer, topicoptions.ReadTopic(topicPath),
+			topicoptions.WithReaderTrace(tracer),
+		)
+		require.NoError(t, err)
+
+		readCtx, firstReaderStopRead := context.WithCancel(ctx)
+		firstReaderReadStopped := make(empty.Chan)
+		go func() {
+			defer close(firstReaderReadStopped)
+
+			for {
+				if readCtx.Err() != nil {
+					return
+				}
+				_, err = firstReader.ReadMessage(readCtx)
+				if readCtx.Err() == nil {
+					require.NoError(t, err)
+				}
+			}
+		}()
+
+		xtest.SpinWaitConditionWithTimeout(t, nil, time.Second, func() bool {
+			return atomic.LoadInt32(&connectedPartitions) == 2
+		})
+
+		readerSecond, err := db.Topic().StartReader(consumer, topicoptions.ReadTopic(topicPath))
+		require.NoError(t, err)
+
+		xtest.SpinWaitConditionWithTimeout(t, nil, time.Second, func() bool {
+			return atomic.LoadInt32(&connectedPartitions) == 1
+		})
+
+		require.NoError(t, readerSecond.Close(ctx))
+
+		xtest.SpinWaitConditionWithTimeout(t, nil, time.Second, func() bool {
+			return atomic.LoadInt32(&connectedPartitions) == 2
+		})
+
+		firstReaderStopRead()
+		xtest.WaitChannelClosed(t, firstReaderReadStopped)
+		require.NoError(t, firstReader.Close(ctx))
+	})
 }
 
 func createCDCFeed(ctx context.Context, t *testing.T, db ydb.Connection) {
@@ -184,7 +268,7 @@ WITH (
 	require.NoError(t, err)
 }
 
-func createReader(t *testing.T, db ydb.Connection, opts ...topicoptions.ReaderOption) *topicreader.Reader {
+func createFeedReader(t *testing.T, db ydb.Connection, opts ...topicoptions.ReaderOption) *topicreader.Reader {
 	topicPath := db.Name() + "/test/feed"
 	reader, err := db.Topic().StartReader("test", []topicoptions.ReadSelector{
 		{
@@ -202,7 +286,7 @@ func createFeedAndReader(
 ) (ydb.Connection, *topicreader.Reader) {
 	db := connect(t)
 	createCDCFeed(ctx, t, db)
-	reader := createReader(t, db, opts...)
+	reader := createFeedReader(t, db, opts...)
 	return db, reader
 }
 
@@ -221,7 +305,7 @@ func sendCDCMessage(ctx context.Context, t *testing.T, db ydb.Connection) {
 	require.NoError(t, err)
 }
 
-func testCtx(t *testing.T) context.Context {
+func testCtx(t testing.TB) context.Context {
 	ctx, cancel := xcontext.WithErrCancel(context.Background())
 	t.Cleanup(func() {
 		cancel(fmt.Errorf("ydb e2e test finished: %v", t.Name()))