+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+* Fixed deadlock in topicreader batcher, while add and read raw server messages
+* Fixed bug in `db.Topic()` with send response to stop partition message
+
 ## v3.32.1
 * Fixed flaky TestTable
 * Renamed topic events in `trace.Details` enum